pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "credit-risk-poc"
version = "0.1.0"
description = "Home Credit anti-fraud feature mining POC"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "PyYAML",
  "scikit-learn",
  "xgboost"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
